refactor(test): rename DummyService to PublicTransitServiceFake
